refactor: moved 'crawlerFactoryFor' into utils.
