+++ lib/src/utils.dart
@@ -1,7 +1,8 @@
 import 'package:annotations/annotations.dart';
+import 'package:dio/dio.dart';
+import 'package:nacht_sources/generated/crawlers.g.dart';
 import 'package:nacht_sources/src/interfaces/crawler.dart';
 import 'package:nacht_sources/src/models/meta.dart';
-import 'package:dio/dio.dart';
 
 /// A crawler factory class used to hold crawler
 /// helper methods
@@ -12,4 +13,14 @@ class CrawlerFactory {
   final Crawler Function(Dio) custom;
 
   const CrawlerFactory(this.meta, this.basic, this.custom);
+}
+
+CrawlerFactory? crawlerFactoryFor(String url, [Dio? client]) {
+  for (final tuple in crawlers) {
+    if (tuple.meta().of(url)) {
+      return tuple;
+    }
+  }
+
+  return null;
 }
